requirements.txt
asgiref==3.4.1
crispy-bootstrap5==0.6
Django==3.2.9
django-crispy-forms==1.14.0
django-dotenv==1.4.2
django-htmx==1.9.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.2
uWSGI==2.0.20
whitenoise==6.0.0
